requirements.txt
pdfplumber
openpyxl
pandas
python-dotenv
pytesseract
pymupdf
pillow
xlrd